Refactor existing logic into a reusable component
